Better handling of dCP
The CP term is cyclic, so introducing bounds and mirroring biases
the way the likelihood works since the endpoints are actually
adjacent.  This removes all the bounding for CP, but we will need
to keep an eye on it to make sure MCMC chains don't diverge to
"infinity and beyond".